move Bridge related Contracts to Bridge namespace
