+++ src/Rector/Architecture/RepositoryAsService/MoveRepositoryFromParentToConstructorRector.php
@@ -8,10 +8,10 @@ use PhpParser\Node;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\Expression;
+use Rector\Bridge\Contract\DoctrineEntityAndRepositoryMapperInterface;
 use Rector\Builder\Class_\VariableInfoFactory;
 use Rector\Builder\ConstructorMethodBuilder;
 use Rector\Builder\PropertyBuilder;
-use Rector\Contract\Bridge\DoctrineEntityAndRepositoryMapperInterface;
 use Rector\Exception\Bridge\RectorProviderException;
 use Rector\Node\Attribute;
 use Rector\Node\NodeFactory;
