+++ src/Rector/Architecture/RepositoryAsService/ServiceLocatorToDIRector.php
@@ -8,8 +8,8 @@ use PhpParser\Node\Expr\ClassConstFetch;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Name;
 use PhpParser\Node\Scalar\String_;
+use Rector\Bridge\Contract\DoctrineEntityAndRepositoryMapperInterface;
 use Rector\Builder\Class_\ClassPropertyCollector;
-use Rector\Contract\Bridge\DoctrineEntityAndRepositoryMapperInterface;
 use Rector\Exception\Bridge\RectorProviderException;
 use Rector\Exception\ShouldNotHappenException;
 use Rector\Naming\PropertyNaming;
