+++ src/Rector/Contrib/Symfony/FrameworkBundle/AbstractToConstructorInjectionRector.php
@@ -7,8 +7,8 @@ use PhpParser\Node\Expr\ClassConstFetch;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Name;
 use PhpParser\Node\Scalar\String_;
+use Rector\Bridge\Contract\ServiceTypeForNameProviderInterface;
 use Rector\Builder\Class_\ClassPropertyCollector;
-use Rector\Contract\Bridge\ServiceTypeForNameProviderInterface;
 use Rector\Naming\PropertyNaming;
 use Rector\Node\Attribute;
 use Rector\Node\PropertyFetchNodeFactory;
