+++ tests/Rector/Contrib/Symfony/HttpKernel/GetterToPropertyRector/Source/rector.yml
@@ -4,4 +4,4 @@ services:
     Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector\Source\DummyProvider: ~
 
     # require for interface autowiring
-    'Rector\Contract\Bridge\ServiceTypeForNameProviderInterface': '@Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector\Source\DummyProvider'
+    'Rector\Bridge\Contract\ServiceTypeForNameProviderInterface': '@Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector\Source\DummyProvider'
